refactor(slider): unify platform implementation
